+++ src/runtime_src/xdp/profile/plugin/opencl/counters/opencl_counters_cb.h
@@ -18,10 +18,12 @@
 #define OPENCL_COUNTERS_CALLBACKS_DOT_H
 
 // These are the functions that are visible when the plugin is dynamically
-//  linked in.  XRT should call them directly.
+//  linked in.  XRT should call them directly.  We use unsigned long long int
+//  so that both Windows and Linux can pass 64-bit values natively via a
+//  C interface.
 
 extern "C"
-void log_function_call_start(const char* functionName, unsigned long int queueAddress, bool isOOO) ;
+void log_function_call_start(const char* functionName, unsigned long long int queueAddress, bool isOOO) ;
 
 extern "C"
 void log_function_call_end(const char* functionName) ;
@@ -29,9 +31,9 @@ void log_function_call_end(const char* functionName) ;
 extern "C"
 void log_kernel_execution(const char* kernelName,
 			  bool isStart,
-			  unsigned long int kernelInstanceAddress,
-			  unsigned long int contextId,
-			  unsigned long int commandQueueId,
+			  unsigned long long int kernelInstanceAddress,
+			  unsigned long long int contextId,
+			  unsigned long long int commandQueueId,
 			  const char* deviceName,
 			  const char* globalWorkSize,
 			  const char* localWorkSize) ;
@@ -43,23 +45,23 @@ void log_compute_unit_execution(const char* cuName,
 				bool isStart) ;
 
 extern "C"
-void counter_action_read(unsigned long int contextId,
-			 unsigned long int numDevices,
+void counter_action_read(unsigned long long int contextId,
+			 unsigned long long int numDevices,
 			 const char* deviceName,
-			 unsigned long int size,
+			 unsigned long long int size,
 			 bool isStart,
 			 bool isP2P,
-			 unsigned long int address,
-			 unsigned long int commandQueueId) ;
+			 unsigned long long int address,
+			 unsigned long long int commandQueueId) ;
 
 extern "C"
-void counter_action_write(unsigned long int contextId,
+void counter_action_write(unsigned long long int contextId,
 			  const char* deviceName,
-			  unsigned long int size,
+			  unsigned long long int size,
 			  bool isStart,
 			  bool isP2P,
-			  unsigned long int address,
-			  unsigned long int commandQueueId) ;
+			  unsigned long long int address,
+			  unsigned long long int commandQueueId) ;
 
 extern "C"
 void counter_mark_objects_released() ;
